AWS S3 Signature V4 - Use the endpoints.json map to get the region
Existing code tried to programatically translate a host/endpoint
to a region for usage in the Signature V4 scope. Unfortunately these
rules may not always apply especially for non-AWS providers. Boto
already has a endpoints.json file which for each service maps a
region to an endpoint. We can make use of this. This additionally
allows other providers to plugin their mappings. The only issue
is that sometimes AWS S3 regions have more than one endpoint so
the additional extra endpoints have to be mapped separately. It's
not too bad though (I hope!).